% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLM.R
\name{myLM}
\alias{myLM}
\title{Linear Regression}
\usage{
myLM(X, Y)
}
\arguments{
\item{X}{an n x p matrix of explanatory variables}

\item{Y}{is an n dimensional vector of responses}
}
\description{
This function does Linear Regression.
}
\keyword{LR}
\keyword{Linear}
\keyword{Regression}
