% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLogistic.R
\name{myLogistic}
\alias{myLogistic}
\title{Logistic Regression}
\usage{
myLogistic(X, Y)
}
\arguments{
\item{X}{an n x p matrix of explanatory variables}

\item{Y}{is an n dimensional vector of responses}
}
\description{
Perform the logistic regression of Y on X
}
\keyword{Logistic}
\keyword{Regression}
