% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPCA.R
\name{myPCA}
\alias{myPCA}
\title{Priciple Composition analysis}
\usage{
myPCA(A, numIter = 1000)
}
\arguments{
\item{A}{Square matrix}
}
\description{
Perform PCA on matrix A using your QR function.
}
\keyword{PCA}
\keyword{QR}
\keyword{decomposition}
