% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myQR.R
\name{myQR}
\alias{myQR}
\title{QR decomposition}
\usage{
myQR(A)
}
\arguments{
\item{A, }{an n x m matrix}
}
\description{
Perform QR decomposition on the matrix A
}
\keyword{QR}
\keyword{decomposition}
